# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  contracts.BatchNotificationRequest:
    properties:
      batch_mode:
        description: 是否批量发送
        type: boolean
      fail_fast:
        description: 遇到错误是否立即停止
        type: boolean
      notifications:
        items:
          $ref: '#/definitions/contracts.NotificationRequest'
        type: array
    required:
    - notifications
    type: object
...
  contracts.DownloadNotificationRequest:
    type: object
...
  contracts.NotificationChannel:
    enum:
    - telegram
    - webhook
    - email
    - system
    type: string
    x-enum-varnames:
    - ChannelTelegram
    - ChannelWebhook
    - ChannelEmail
    - ChannelSystem
  contracts.NotificationLevel:
    enum:
    - info
    - warning
    - error
    - success
    type: string
    x-enum-varnames:
    - NotificationLevelInfo
    - NotificationLevelWarning
    - NotificationLevelError
    - NotificationLevelSuccess
  contracts.NotificationRequest:
    properties:
      channel:
        $ref: '#/definitions/contracts.NotificationChannel'
      data:
        additionalProperties: true
        type: object
      level:
        $ref: '#/definitions/contracts.NotificationLevel'
      message:
        type: string
      priority:
        type: integer
      target_id:
        description: 如Telegram chat_id
        type: string
      template:
        type: string
      title:
        type: string
    required:
    - channel
    - level
    - message
    - title
    type: object
...
  contracts.SystemNotificationRequest:
    properties:
      component:
        description: aria2, alist, scheduler, etc.
        type: string
      details:
        additionalProperties: true
        type: object
      event:
        description: startup, shutdown, error, etc.
        type: string
      level:
        $ref: '#/definitions/contracts.NotificationLevel'
      message:
        type: string
      metrics:
        additionalProperties: true
        type: object
    required:
    - component
    - event
    - level
    - message
    type: object
...
  contracts.TaskNotificationRequest:
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 基于Gin框架的Alist和Aria2下载管理服务
  license:
    name: MIT
    url: https://github.com/easayliu/alist-aria2-download/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Alist Aria2 Download API
  version: "1.0"
paths:
...
  /notifications/batch:
    post:
      consumes:
      - application/json
      description: 批量发送多个通知消息
      parameters:
      - description: 批量通知请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/contracts.BatchNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 批量通知发送结果
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 批量发送通知
      tags:
      - 通知管理
  /notifications/config:
    get:
      description: 获取通知系统的配置信息
      produces:
      - application/json
      responses:
        "200":
          description: 通知配置信息
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取通知配置
      tags:
      - 通知管理
  /notifications/download-complete:
    post:
      consumes:
      - application/json
      description: 发送下载完成通知
      parameters:
      - description: 下载通知请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/contracts.DownloadNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 通知发送成功
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 下载完成通知
      tags:
      - 通知管理
  /notifications/download-failed:
    post:
      consumes:
      - application/json
      description: 发送下载失败通知
      parameters:
      - description: 下载通知请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/contracts.DownloadNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 通知发送成功
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 下载失败通知
      tags:
      - 通知管理
  /notifications/history:
    get:
      description: 获取历史通知记录列表
      parameters:
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 通知历史列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取通知历史
      tags:
      - 通知管理
  /notifications/send:
    post:
      consumes:
      - application/json
      description: 发送单个通知消息
      parameters:
      - description: 通知请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/contracts.NotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 通知发送成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 发送通知
      tags:
      - 通知管理
  /notifications/stats:
    get:
      description: 获取通知系统的统计信息
      produces:
      - application/json
      responses:
        "200":
          description: 通知统计信息
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取通知统计
      tags:
      - 通知管理
  /notifications/system-event:
    post:
      consumes:
      - application/json
      description: 发送系统事件通知
      parameters:
      - description: 系统通知请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/contracts.SystemNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 通知发送成功
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 系统事件通知
      tags:
      - 通知管理
  /notifications/task-complete:
    post:
      consumes:
      - application/json
      description: 发送任务完成通知
      parameters:
      - description: 任务通知请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/contracts.TaskNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 通知发送成功
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 任务完成通知
      tags:
      - 通知管理
  /notifications/task-failed:
    post:
      consumes:
      - application/json
      description: 发送任务失败通知
      parameters:
      - description: 任务通知请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/contracts.TaskNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 通知发送成功
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 任务失败通知
      tags:
      - 通知管理
...
schemes:
- http
- https
swagger: "2.0"